requirements.txt
aiogram
python-dotenv
openai==0.28